refactor: Drop mergeStrategy from the api surface + expose helpers
This is more flexible and it feels like better composition.

Closes #61.
